Cardano.Protocol.Socket.Type

Documentation

type Tip = Block

Tip of the block chain type (used by node protocols).

newtype BlockId

The node protocols require a block header type.

Constructors

BlockId

Fields

getBlockId :: ByteString

blockId :: Block -> BlockId

A hash of the block's contents.

maximumMiniProtocolLimits :: MiniProtocolLimits

Limits for the protocols we use.

nodeToClientVersion :: NodeToClientVersion

Protocol versions

cfgNetworkMagic :: NetworkMagic

A temporary definition of the protocol version. This will be moved as an argument to the client connection function in a future PR (the network magic number matches the one in the test net created by scripts)

cfgNetworkId :: NetworkId

nodeToClientVersionData :: NodeToClientVersionData

doNothingInitiatorProtocol :: MonadTimer m => RunMiniProtocol 'InitiatorMode ByteString m a Void

A protocol client that will never leave the initial state.

doNothingResponderProtocol :: MonadTimer m => RunMiniProtocol 'ResponderMode ByteString m Void a

type Offset = Integer

epochSlots :: EpochSlots

Boilerplate codecs used for protocol serialisation.

The number of epochSlots is specific to each blockchain instance. This value is what the cardano main and testnet uses. Only applies to the Byron era.

codecVersion :: BlockNodeToClientVersion (CardanoBlock StandardCrypto)

codecConfig :: CodecConfig (CardanoBlock StandardCrypto)

nodeToClientCodecs :: forall m. MonadST m => ClientCodecs (CardanoBlock StandardCrypto) m

chainSyncCodec :: forall block. (Serialise block, Serialise (HeaderHash block)) => Codec (ChainSync block (Point block) Tip) DeserialiseFailure IO ByteString

These codecs are currently used in the mock nodes and will probably soon get removed as the mock nodes are phased out.

txSubmissionCodec :: Codec (LocalTxSubmission (Tx BabbageEra) String) DeserialiseFailure IO ByteString

Orphan instances

NoThunks TxId

Instance details

Methods

noThunks :: Context -> TxId -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> TxId -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy TxId -> String

NoThunks BuiltinByteString

Instance details

Methods

noThunks :: Context -> BuiltinByteString -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> BuiltinByteString -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy BuiltinByteString -> String

StandardHash Block

Instance details

StandardHash (Tx BabbageEra)

Instance details

ShowProxy Char

Instance details

Methods

showProxy :: Proxy Char -> String

ShowProxy OnChainTx

Instance details

Methods

showProxy :: Proxy OnChainTx -> String

ShowProxy a => ShowProxy ([a] :: Type)

Instance details

Methods

showProxy :: Proxy [a] -> String

ShowProxy (Tx BabbageEra)

Instance details

Methods

showProxy :: Proxy (Tx BabbageEra) -> String