Plutus.ChainIndex.Emulator.DiskState

Description

The disk state is the part of the chain index that is kept on disk. This module defines an in-memory implementation of the disk state which can be used in the emulator.

Documentation

data DiskState

Data that we keep on disk. (This type is used for testing only - we need other structures for the disk-backed storage)

dataMap :: Lens' DiskState (Map DatumHash Datum)

scriptMap :: Lens' DiskState (Map ScriptHash (Versioned Script))

redeemerMap :: Lens' DiskState (Map RedeemerHash Redeemer)

txMap :: Lens' DiskState (Map TxId ChainIndexTx)

addressMap :: Lens' DiskState CredentialMap

assetClassMap :: Lens' DiskState AssetClassMap

fromTx :: ChainIndexTx -> DiskState

The data we store on disk for a given ChainIndexTx

data CredentialMap

Set of transaction output references for each address.

unCredentialMap :: Iso' CredentialMap (Map Credential (Set TxOutRef))

data AssetClassMap

Set of transaction output references for each asset class.

unAssetClassMap :: Iso' AssetClassMap (Map AssetId (Set TxOutRef))

diagnostics :: DiskState -> Diagnostics