Plutus.ChainIndex.Tx

Description

The chain index' version of a transaction

Documentation

data ChainIndexTx

Constructors

ChainIndexTx

Fields

_citxTxId :: TxId

The id of this transaction.

_citxInputs :: [TxOutRef]

The inputs to this transaction.

_citxOutputs :: ChainIndexTxOutputs

The outputs of this transaction, ordered so they can be referenced by index.

_citxValidRange :: !SlotRange

The SlotRange during which this transaction may be validated.

_citxData :: Map DatumHash Datum

Datum objects recorded on this transaction.

_citxRedeemers :: Redeemers

Redeemers of the minting scripts.

_citxScripts :: Map ScriptHash (Versioned Script)

The scripts (validator, stake validator or minting) part of cardano tx.

_citxCardanoTx :: Maybe CardanoTx

The full Cardano API tx which was used to populate the rest of the ChainIndexTx fields. Useful because ChainIndexTx doesn't have all the details of the tx, so we keep it as a safety net. Might be Nothing if we are in the emulator.

data ChainIndexTxOutputs

List of outputs of a transaction. There is only an optional collateral output if the transaction is invalid.

Constructors

InvalidTx (Maybe ChainIndexTxOut)
The transaction is invalid so there is maybe a collateral output.

ValidTx [ChainIndexTxOut]

data ChainIndexTxOut

Constructors

ChainIndexTxOut

Fields

citoAddress :: CardanoAddress

We can't use AddressInAnyEra here because of missing FromJson instance for Byron era

citoValue :: Value

citoDatum :: OutputDatum

citoRefScript :: ReferenceScript

data ReferenceScript

Constructors

ReferenceScriptNone

ReferenceScriptInAnyLang ScriptInAnyLang

data Address

Constructors

Address

Fields

addressCredential :: Credential

addressStakingCredential :: Maybe StakingCredential

data OutputDatum

Constructors

NoOutputDatum

OutputDatumHash DatumHash

OutputDatum Datum

newtype Value

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Integer)

fromOnChainTx :: OnChainTx -> ChainIndexTx

Convert a OnChainTx to a ChainIndexTx. An invalid OnChainTx will not produce any ChainIndexTx outputs and the collateral inputs of the OnChainTx will be the inputs of the ChainIndexTx.

txOuts :: ChainIndexTx -> [ChainIndexTxOut]

Get tx outputs from tx.

txOutRefs :: ChainIndexTx -> [TxOutRef]

Get tx output references from tx.

txOutsWithRef :: ChainIndexTx -> [(ChainIndexTxOut, TxOutRef)]

Get tx output references and tx outputs from tx.

txOutRefMap :: ChainIndexTx -> Map TxOutRef (ChainIndexTxOut, ChainIndexTx)

Get Map of tx outputs references to tx.

txOutRefMapForAddr :: CardanoAddress -> ChainIndexTx -> Map TxOutRef (ChainIndexTxOut, ChainIndexTx)

Get Map of tx outputs from tx for a specific address.

txRedeemersWithHash :: ChainIndexTx -> Map RedeemerHash Redeemer

validityFromChainIndex :: ChainIndexTx -> TxValidity

Lenses

citxTxId :: Lens' ChainIndexTx TxId

citxInputs :: Lens' ChainIndexTx [TxOutRef]

citxOutputs :: Lens' ChainIndexTx ChainIndexTxOutputs

citxValidRange :: Lens' ChainIndexTx SlotRange

citxData :: Lens' ChainIndexTx (Map DatumHash Datum)

citxRedeemers :: Lens' ChainIndexTx Redeemers

citxScripts :: Lens' ChainIndexTx (Map ScriptHash (Versioned Script))

citxCardanoTx :: Lens' ChainIndexTx (Maybe CardanoTx)

_InvalidTx :: Prism' ChainIndexTxOutputs (Maybe ChainIndexTxOut)

_ValidTx :: Prism' ChainIndexTxOutputs [ChainIndexTxOut]