Plutus.ChainIndex.TxIdState

Documentation

increaseDepth :: TxStatus -> TxStatus

Increase the depth of a tentatively confirmed transaction

initialStatus :: OnChainTx -> TxStatus

The TxStatus of a transaction right after it was added to the chain

transactionStatus :: BlockNumber -> TxIdState -> TxId -> Either TxStatusFailure TxStatus

Given the current block, compute the status for the given transaction by checking to see if it has been deleted.

fromTx :: BlockNumber -> ChainIndexTx -> TxIdState

fromBlock :: Tip -> [ChainIndexTx] -> UtxoState TxIdState

rollback :: Point -> UtxoIndex TxIdState -> Either RollbackFailed (RollbackResult TxIdState)

chainConstant :: Depth

The depth (in blocks) after which a transaction cannot be rolled back anymore

dropOlder :: Monoid a => BlockNumber -> UtxoIndex a -> UtxoIndex a

Drop everything older than BlockNumber in the index.