Plutus.ChainIndex.TxOutBalance

Documentation

transactionOutputStatus

Arguments

:: BlockNumber
Current block number for inspecting the state of the transaction output

-> TxIdState
Information on the state of a transaction. Needed for determining its status.

-> TxOutBalance
Balance of spent and unspent transaction outputs.

-> TxOutRef
Target transaction output for inspecting its state.

-> Either TxStatusFailure TxOutStatus

Given the current block, compute the status for the given transaction output by getting the state of the transaction that produced it and checking if the output is spent or unspent.

transactionOutputState :: TxOutBalance -> TxOutRef -> Maybe TxOutState

fromTx :: ChainIndexTx -> TxOutBalance

isUnspentOutput :: TxOutRef -> UtxoState TxOutBalance -> Bool

Whether a TxOutRef is a member of the UTXO set (ie. unspent)

unspentOutputs :: UtxoState TxOutBalance -> Set TxOutRef

The UTXO set

isSpentOutput :: TxOutRef -> UtxoState TxOutBalance -> Bool

Whether a TxOutRef is a member of the spent tx output set.

spentOutputs :: UtxoState TxOutBalance -> Set TxOutRef

The spent output set

fromBlock :: Tip -> [ChainIndexTx] -> UtxoState TxOutBalance

UtxoIndex for a single block

rollback :: Point -> UtxoIndex TxOutBalance -> Either RollbackFailed (RollbackResult TxOutBalance)

Perform a rollback on the utxo index