Plutus.ChainIndex.UtxoState

Description

The UTXO state, kept in memory by the chain index.

Documentation

data UtxoState a

UTXO / ledger state, kept in memory. We are only interested in the UTXO set, everything else is stored on disk. This is OK because we don't need to validate transactions when they come in.

Constructors

UtxoState

Fields

_usTxUtxoData :: a

_usTip :: Tip

Tip of our chain sync client

usTxUtxoData :: forall a a. Lens (UtxoState a) (UtxoState a) a a

usTip :: forall a. Lens' (UtxoState a) Tip

type UtxoIndex a = FingerTree (BlockCount, UtxoState a) (UtxoState a)

utxoState :: Monoid a => UtxoIndex a -> UtxoState a

utxoBlockCount :: Monoid a => UtxoIndex a -> Int

tip :: UtxoState a -> Tip

viewTip :: Monoid a => UtxoIndex a -> Tip

pointLessThanTip :: Point -> Tip -> Bool

Is the given point earlier than the provided tip. Yes, if the point is the genersis point, no if the tip is the genesis point, otherwise, just compare the slots.

Extending the UTXO index

data InsertUtxoPosition

Outcome of inserting a UtxoState into the utxo index

Constructors

InsertAtEnd
The utxo state was added to the end. Returns the new index

InsertBeforeEnd
The utxo state was added somewhere before the end. Returns the new index and the tip

data InsertUtxoSuccess a

Constructors

InsertUtxoSuccess

Fields

newIndex :: UtxoIndex a

insertPosition :: InsertUtxoPosition

data InsertUtxoFailed

UTXO state could not be inserted into the chain index

Constructors

DuplicateBlock Tip
Insertion failed as there was already a block with the given number

InsertUtxoNoTip
The _usTip field of the argument was 'Last Nothing'

insert :: (Monoid a, Eq a) => UtxoState a -> UtxoIndex a -> Either InsertUtxoFailed (InsertUtxoSuccess a)

Insert a UtxoState into the index

Rollbacks

data RollbackFailed

Reason why the rollback operation failed

Constructors

RollbackNoTip
Rollback failed because the utxo index had no tip (not synchronised)

TipMismatch
Unable to roll back to expectedTip because the tip at that position was different

Fields

foundTip :: Tip

targetPoint :: Point

OldPointNotFound Point
Unable to find the old tip

data RollbackResult a

Constructors

RollbackResult

Fields

newTip :: Tip

rolledBackIndex :: UtxoIndex a

rollbackWith

Arguments

:: Monoid a

=> (UtxoIndex a -> UtxoIndex a -> UtxoIndex a)
Calculate the new index given the index before and the index after the rollback point.

-> Point

-> UtxoIndex a

-> Either RollbackFailed (RollbackResult a)

Perform a rollback on the utxo index, with a callback to calculate the new index.

Limit the UTXO index size

data ReduceBlockCountResult a

Constructors

BlockCountNotReduced

ReduceBlockCountResult

Fields

reducedIndex :: UtxoIndex a

combinedState :: UtxoState a

reduceBlockCount :: Monoid a => Depth -> UtxoIndex a -> ReduceBlockCountResult a

Reduce the number of UtxoStates. The given number is the minimum, the index is reduced when it larger than twice that size. The new index is prefixed with one UtxoState that contains the combined state of the removed UtxoStates.

newtype BlockCount

Constructors

BlockCount

Fields

getBlockCount :: Int