Cardano.Protocol.Socket.Client

Documentation

data ChainSyncHandle event

Constructors

ChainSyncHandle

Fields

cshCurrentSlot :: IO Slot

cshHandler :: event -> Slot -> IO ()

data ChainSyncEvent

Constructors

Resume !ChainPoint

RollForward !(BlockInMode CardanoMode) !ChainTip

RollBackward !ChainPoint !ChainTip

type ChainSyncCallback = ChainSyncEvent -> Slot -> IO ()

The Slot parameter here represents the current slot as computed from the current time. There is also the slot where the block was published, which is available from the ChainSyncEvent.

Currently we are using this current slot everywhere, which is why I leave it here, as a parameter.

data ClientMsg

Constructors

Disconnected Text

Resumed Point

RolledForward Tip

RolledBackward Point

getCurrentSlot :: forall block. ChainSyncHandle block -> IO Slot

runChainSync' :: FilePath -> SlotConfig -> NetworkId -> [ChainPoint] -> IO (ChainSyncHandle ChainSyncEvent)

Run the chain sync protocol to get access to the current slot number.

runChainSync :: FilePath -> Trace IO ClientMsg -> SlotConfig -> NetworkId -> [ChainPoint] -> (ChainSyncEvent -> IO ()) -> IO (ChainSyncHandle ChainSyncEvent)

chainSyncClient :: Trace IO ClientMsg -> SlotConfig -> [ChainPoint] -> ChainSyncCallback -> ChainSyncClient (BlockInMode CardanoMode) ChainPoint ChainTip IO ()

The client updates the application state when the protocol state changes.