Plutus.ChainIndex

Documentation

runChainIndexEffects :: RunRequirements -> Eff '[ChainIndexQueryEffect, ChainIndexControlEffect, BeamEffect Sqlite] a -> IO (Either ChainIndexError a)

Run the chain index effects.

handleChainIndexEffects :: (LastMember IO effs, Member (LogMsg ChainIndexLog) effs) => RunRequirements -> Eff (ChainIndexQueryEffect ': (ChainIndexControlEffect ': (BeamEffect Sqlite ': effs))) a -> Eff effs (Either ChainIndexError a)

Handle the chain index effects from the set of all effects.

data RunRequirements

The required arguments to run the chain index effects.

Constructors

RunRequirements

Fields

trace :: Trace IO (PrettyObject ChainIndexLog)

stateTVar :: TVar ChainIndexState

pool :: Pool Connection

securityParam :: Int

newtype Value

A cryptocurrency value. This is a map from CurrencySymbols to a quantity of that currency.

Operations on currencies are usually implemented pointwise. That is, we apply the operation to the quantities for each currency in turn. So when we add two Values the resulting Value has, for each currency, the sum of the quantities of that particular currency in the argument Value. The effect of this is that the currencies in the Value are "independent", and are operated on separately.

Whenever we need to get the quantity of a currency in a Value where there is no explicit quantity of that currency in the Value, then the quantity is taken to be zero.

See note [Currencies] for more details.

Constructors

Value

Fields

getValue :: Map CurrencySymbol (Map TokenName Integer)

pageOf

Arguments

:: Eq a

=> PageQuery a
Pagination query parameters.

-> Set a

-> Page a

Given a Set, request the Page with the given PageQuery.

data PageQuery a

Query parameters for pagination.

Constructors

PageQuery

Fields

pageQuerySize :: PageSize

Number of items per page.

pageQueryLastItem :: Maybe a

Last item of the queried page.

newtype PageSize

Constructors

PageSize

Fields

getPageSize :: Natural

data Page a

Part of a collection.

Constructors

Page

Fields

currentPageQuery :: PageQuery a

The PageQuery which was used to request this Page.

nextPageQuery :: Maybe (PageQuery a)

The PageQuery to use to request the next Page. Nothing if we requested the last page.

pageItems :: [a]

Items in the current Page.

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

data Address

Address with two kinds of credentials, normal and staking.

Constructors

Address

Fields

addressCredential :: Credential

addressStakingCredential :: Maybe StakingCredential

data OutputDatum

The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").

Constructors

NoOutputDatum

OutputDatumHash DatumHash

OutputDatum Datum

data ChainIndexTxOutputs

List of outputs of a transaction. There is only an optional collateral output if the transaction is invalid.

Constructors

InvalidTx (Maybe ChainIndexTxOut)
The transaction is invalid so there is maybe a collateral output.

ValidTx [ChainIndexTxOut]

data ChainIndexTxOut

Constructors

ChainIndexTxOut

Fields

citoAddress :: CardanoAddress

We can't use AddressInAnyEra here because of missing FromJson instance for Byron era

citoValue :: Value

citoDatum :: OutputDatum

citoRefScript :: ReferenceScript

data ReferenceScript

Constructors

ReferenceScriptNone

ReferenceScriptInAnyLang ScriptInAnyLang

fromReferenceScript :: ReferenceScript -> Maybe (Versioned Script)

data ChainIndexTx

Constructors

ChainIndexTx

Fields

_citxTxId :: TxId

The id of this transaction.

_citxInputs :: [TxIn]

The inputs to this transaction.

_citxOutputs :: ChainIndexTxOutputs

The outputs of this transaction, ordered so they can be referenced by index.

_citxValidRange :: !SlotRange

The SlotRange during which this transaction may be validated.

_citxData :: Map DatumHash Datum

Datum objects recorded on this transaction.

_citxRedeemers :: Redeemers

Redeemers of the minting scripts.

_citxScripts :: Map ScriptHash (Versioned Script)

The scripts (validator, stake validator or minting) part of cardano tx.

_citxCardanoTx :: Maybe SomeCardanoApiTx

The full Cardano API tx which was used to populate the rest of the ChainIndexTx fields. Useful because ChainIndexTx doesn't have all the details of the tx, so we keep it as a safety net. Might be Nothing if we are in the emulator.

_InvalidTx :: Prism' ChainIndexTxOutputs (Maybe ChainIndexTxOut)

_ValidTx :: Prism' ChainIndexTxOutputs [ChainIndexTxOut]

data Tip

The tip of the chain index.

Constructors

TipAtGenesis

Tip

Fields

tipSlot :: Slot

Last slot

tipBlockId :: BlockId

Last block ID

tipBlockNo :: BlockNumber

Last block number

newtype BlockNumber

Constructors

BlockNumber

Fields

unBlockNumber :: Word64

citxCardanoTx :: Lens' ChainIndexTx (Maybe SomeCardanoApiTx)

citxData :: Lens' ChainIndexTx (Map DatumHash Datum)

citxInputs :: Lens' ChainIndexTx [TxIn]

citxOutputs :: Lens' ChainIndexTx ChainIndexTxOutputs

citxRedeemers :: Lens' ChainIndexTx Redeemers

citxScripts :: Lens' ChainIndexTx (Map ScriptHash (Versioned Script))

citxTxId :: Lens' ChainIndexTx TxId

citxValidRange :: Lens' ChainIndexTx SlotRange

blockId :: Block -> BlockId

Compute a hash of the block's contents.

data Point

When performing a rollback the chain sync protocol does not provide a block number where to resume from.

Constructors

PointAtGenesis

Point

Fields

pointSlot :: Slot

Slot number

pointBlockId :: BlockId

Block number

data TxOutBalance

The effect of a transaction (or a number of them) on the tx output set.

Constructors

TxOutBalance

Fields

_tobUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tobSpentOutputs :: Map TxOutRef TxId

Outputs spent by the transaction(s) along with the tx id that spent it

data TxConfirmedState

Constructors

TxConfirmedState

Fields

timesConfirmed :: Sum Int

blockAdded :: Last BlockNumber

validity :: Last TxValidity

data TxIdState

Constructors

TxIdState

Fields

txnsConfirmed :: Map TxId TxConfirmedState

Number of times this transaction has been added as well as other necessary metadata.

txnsDeleted :: Map TxId (Sum Int)

Number of times this transaction has been deleted.

data TxStatusFailure

Datatype returned when we couldn't get the state of a tx or a tx output.

Constructors

TxIdStateInvalid BlockNumber TxId TxIdState
We couldn't return the status because the TxIdState was in a ... state ... that we didn't know how to decode in transactionStatus.

TxOutBalanceStateInvalid BlockNumber TxOutRef TxOutBalance
We couldn't return the status because the TxOutBalance does not contain the target tx output.

InvalidRollbackAttempt BlockNumber TxId TxIdState

data Diagnostics

Constructors

Diagnostics

Fields

numTransactions :: Integer

numScripts :: Integer

numAddresses :: Integer

numAssetClasses :: Integer

numUnspentOutputs :: Int

numUnmatchedInputs :: Int

someTransactions :: [TxId]

unspentTxOuts :: [ChainIndexTxOut]

data TxOutState

Constructors

Spent TxId

Unspent

type TxOutStatus = RollbackState TxOutState

data RollbackState a

The rollback state of a Cardano transaction

Constructors

Unknown
The transaction is not on the chain. That's all we can say.

TentativelyConfirmed Depth TxValidity a
The transaction is on the chain, n blocks deep. It can still be rolled back.

Committed TxValidity a
The transaction is on the chain. It cannot be rolled back anymore.

type TxStatus = RollbackState ()

The status of a Cardano transaction

newtype Depth

How many blocks deep the tx is on the chain

Constructors

Depth

Fields

unDepth :: Int

data TxValidity

Validity of a transaction that has been added to the ledger

Constructors

TxValid

TxInvalid

UnknownValidity

_PointAtGenesis :: Prism' Point ()

_Point :: Prism' Point (Slot, BlockId)

tipAsPoint :: Tip -> Point

pointsToTip :: Point -> Tip -> Bool

txOutStatusTxOutState :: TxOutStatus -> Maybe TxOutState

Maybe extract the TxOutState (Spent or Unspent) of a TxOutStatus.

liftTxOutStatus :: TxOutStatus -> TxStatus

Converts a TxOutStatus to a TxStatus. Possible since a transaction output belongs to a transaction.

Note, however, that we can't convert a TxStatus to a TxOutStatus.

data TxUtxoBalance

The effect of a transaction (or a number of them) on the utxo set.

Constructors

TxUtxoBalance

Fields

_tubUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tubUnmatchedSpentInputs :: Set TxOutRef

Outputs spent by the transaction(s) that have no matching unspent output

tobSpentOutputs :: Lens' TxOutBalance (Map TxOutRef TxId)

tobUnspentOutputs :: Lens' TxOutBalance (Set TxOutRef)

data ChainSyncBlock

A block of transactions to be synced.

Constructors

Block

Fields

blockTip :: Tip

blockTxs :: [(ChainIndexTx, TxProcessOption)]

newtype TxProcessOption

User-customizable options to process a transaction. See #73 for more motivations.

Constructors

TxProcessOption

Fields

tpoStoreTx :: Bool

Should the chain index store this transaction or not. If not, only handle the UTXOs. This, for example, allows applications to skip unwanted pre-Alonzo transactions.

tubUnmatchedSpentInputs :: Lens' TxUtxoBalance (Set TxOutRef)

tubUnspentOutputs :: Lens' TxUtxoBalance (Set TxOutRef)

data RollbackFailed

Reason why the rollback operation failed

Constructors

RollbackNoTip
Rollback failed because the utxo index had no tip (not synchronised)

TipMismatch
Unable to roll back to expectedTip because the tip at that position was different

Fields

foundTip :: Tip

targetPoint :: Point

OldPointNotFound Point
Unable to find the old tip

data InsertUtxoFailed

UTXO state could not be inserted into the chain index

Constructors

DuplicateBlock Tip
Insertion failed as there was already a block with the given number

InsertUtxoNoTip
The _usTip field of the argument was 'Last Nothing'

data ChainIndexError

Constructors

InsertionFailed InsertUtxoFailed

RollbackFailed RollbackFailed

ResumeNotSupported

QueryFailedNoTip
Query failed because the chain index does not have a tip (not synchronised with node)

BeamEffectError BeamError

txOuts :: ChainIndexTx -> [ChainIndexTxOut]

Get tx outputs from tx.

txOutRefs :: ChainIndexTx -> [TxOutRef]

Get tx output references from tx.

txOutsWithRef :: ChainIndexTx -> [(ChainIndexTxOut, TxOutRef)]

Get tx output references and tx outputs from tx.

txOutRefMap :: ChainIndexTx -> Map TxOutRef (ChainIndexTxOut, ChainIndexTx)

Get Map of tx outputs references to tx.

txOutRefMapForAddr :: CardanoAddress -> ChainIndexTx -> Map TxOutRef (ChainIndexTxOut, ChainIndexTx)

Get Map of tx outputs from tx for a specific address.

validityFromChainIndex :: ChainIndexTx -> TxValidity

fromOnChainTx :: OnChainTx -> ChainIndexTx

Convert a OnChainTx to a ChainIndexTx. An invalid OnChainTx will not produce any ChainIndexTx outputs and the collateral inputs of the OnChainTx will be the inputs of the ChainIndexTx.

txRedeemersWithHash :: ChainIndexTx -> Map RedeemerHash Redeemer

data ChainIndexQueryEffect r where

Constructors

DatumFromHash :: DatumHash -> ChainIndexQueryEffect (Maybe Datum)
Get the datum from a datum hash (if available)

ValidatorFromHash :: ValidatorHash -> ChainIndexQueryEffect (Maybe (Versioned Validator))
Get the validator from a validator hash (if available)

MintingPolicyFromHash :: MintingPolicyHash -> ChainIndexQueryEffect (Maybe (Versioned MintingPolicy))
Get the monetary policy from an MPS hash (if available)

RedeemerFromHash :: RedeemerHash -> ChainIndexQueryEffect (Maybe Redeemer)
Get the redeemer from a redeemer hash (if available)

StakeValidatorFromHash :: StakeValidatorHash -> ChainIndexQueryEffect (Maybe (Versioned StakeValidator))
Get the stake validator from a stake validator hash (if available)

UnspentTxOutFromRef :: TxOutRef -> ChainIndexQueryEffect (Maybe DecoratedTxOut)
Get the TxOut from a TxOutRef (if available)

TxOutFromRef :: TxOutRef -> ChainIndexQueryEffect (Maybe DecoratedTxOut)
Get the TxOut from a TxOutRef (if available)

TxFromTxId :: TxId -> ChainIndexQueryEffect (Maybe ChainIndexTx)
Get the transaction for a tx ID

UtxoSetMembership :: TxOutRef -> ChainIndexQueryEffect IsUtxoResponse
Whether a tx output is part of the UTXO set

UtxoSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect UtxosResponse
Unspent outputs located at addresses with the given credential.

UnspentTxOutSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect (QueryResponse [(TxOutRef, DecoratedTxOut)])
Get the unspent txouts located at an address This is to avoid multiple queries from chain-index when using utxosAt

DatumsAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect (QueryResponse [Datum])
get the datums located at addresses with the given credential.

UtxoSetWithCurrency :: PageQuery TxOutRef -> AssetClass -> ChainIndexQueryEffect UtxosResponse
Unspent outputs containing a specific currency (AssetClass).

Note that requesting unspent outputs containing Ada should not return anything, as this request will always return all unspent outputs.

TxsFromTxIds :: [TxId] -> ChainIndexQueryEffect [ChainIndexTx]
Get the transactions for a list of tx IDs.

TxoSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect TxosResponse
Outputs located at addresses with the given credential.

GetTip :: ChainIndexQueryEffect Tip
Get the tip of the chain index

data ChainIndexControlEffect r where

Constructors

AppendBlocks :: [ChainSyncBlock] -> ChainIndexControlEffect ()
Add new blocks to the chain index.

Rollback :: Point -> ChainIndexControlEffect ()
Roll back to a previous state (previous tip)

ResumeSync :: Point -> ChainIndexControlEffect ()
Resume syncing from a certain point

CollectGarbage :: ChainIndexControlEffect ()
Delete all data that is not covered by current UTxOs.

GetDiagnostics :: ChainIndexControlEffect Diagnostics

datumFromHash :: forall effs. Member ChainIndexQueryEffect effs => DatumHash -> Eff effs (Maybe Datum)

validatorFromHash :: forall effs. Member ChainIndexQueryEffect effs => ValidatorHash -> Eff effs (Maybe (Versioned Validator))

mintingPolicyFromHash :: forall effs. Member ChainIndexQueryEffect effs => MintingPolicyHash -> Eff effs (Maybe (Versioned MintingPolicy))

redeemerFromHash :: forall effs. Member ChainIndexQueryEffect effs => RedeemerHash -> Eff effs (Maybe Redeemer)

stakeValidatorFromHash :: forall effs. Member ChainIndexQueryEffect effs => StakeValidatorHash -> Eff effs (Maybe (Versioned StakeValidator))

unspentTxOutFromRef :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs (Maybe DecoratedTxOut)

txOutFromRef :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs (Maybe DecoratedTxOut)

txFromTxId :: forall effs. Member ChainIndexQueryEffect effs => TxId -> Eff effs (Maybe ChainIndexTx)

utxoSetMembership :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs IsUtxoResponse

utxoSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs UtxosResponse

unspentTxOutSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs (QueryResponse [(TxOutRef, DecoratedTxOut)])

datumsAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs (QueryResponse [Datum])

utxoSetWithCurrency :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> AssetClass -> Eff effs UtxosResponse

txsFromTxIds :: forall effs. Member ChainIndexQueryEffect effs => [TxId] -> Eff effs [ChainIndexTx]

txoSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs TxosResponse

getTip :: forall effs. Member ChainIndexQueryEffect effs => Eff effs Tip

appendBlocks :: forall effs. Member ChainIndexControlEffect effs => [ChainSyncBlock] -> Eff effs ()

rollback :: forall effs. Member ChainIndexControlEffect effs => Point -> Eff effs ()

resumeSync :: forall effs. Member ChainIndexControlEffect effs => Point -> Eff effs ()

collectGarbage :: forall effs. Member ChainIndexControlEffect effs => Eff effs ()

getDiagnostics :: forall effs. Member ChainIndexControlEffect effs => Eff effs Diagnostics

data InsertUtxoPosition

Outcome of inserting a UtxoState into the utxo index

Constructors

InsertAtEnd
The utxo state was added to the end. Returns the new index

InsertBeforeEnd
The utxo state was added somewhere before the end. Returns the new index and the tip

data ChainIndexLog

Constructors

InsertionSuccess Tip InsertUtxoPosition

ConversionFailed FromCardanoError

RollbackSuccess Tip

Err ChainIndexError

TxNotFound TxId

TxOutNotFound TxOutRef

TipIsGenesis

NoDatumScriptAddr ChainIndexTxOut

BeamLogItem BeamLog

data UtxoState a

UTXO / ledger state, kept in memory. We are only interested in the UTXO set, everything else is stored on disk. This is OK because we don't need to validate transactions when they come in.

Constructors

UtxoState

Fields

_usTxUtxoData :: a

_usTip :: Tip

Tip of our chain sync client

data ReduceBlockCountResult a

Constructors

BlockCountNotReduced

ReduceBlockCountResult

Fields

reducedIndex :: UtxoIndex a

combinedState :: UtxoState a

data RollbackResult a

Constructors

RollbackResult

Fields

newTip :: Tip

rolledBackIndex :: UtxoIndex a

data InsertUtxoSuccess a

Constructors

InsertUtxoSuccess

Fields

newIndex :: UtxoIndex a

insertPosition :: InsertUtxoPosition

type UtxoIndex a = FingerTree (BlockCount, UtxoState a) (UtxoState a)

newtype BlockCount

Constructors

BlockCount

Fields

getBlockCount :: Int

usTip :: forall a. Lens' (UtxoState a) Tip

usTxUtxoData :: forall a a. Lens (UtxoState a) (UtxoState a) a a

utxoState :: Monoid a => UtxoIndex a -> UtxoState a

utxoBlockCount :: Monoid a => UtxoIndex a -> Int

tip :: UtxoState a -> Tip

viewTip :: Monoid a => UtxoIndex a -> Tip

insert :: (Monoid a, Eq a) => UtxoState a -> UtxoIndex a -> Either InsertUtxoFailed (InsertUtxoSuccess a)

Insert a UtxoState into the index

rollbackWith

Arguments

:: Monoid a

=> (UtxoIndex a -> UtxoIndex a -> UtxoIndex a)
Calculate the new index given the index before and the index after the rollback point.

-> Point

-> UtxoIndex a

-> Either RollbackFailed (RollbackResult a)

Perform a rollback on the utxo index, with a callback to calculate the new index.

reduceBlockCount :: Monoid a => Depth -> UtxoIndex a -> ReduceBlockCountResult a

Reduce the number of UtxoStates. The given number is the minimum, the index is reduced when it larger than twice that size. The new index is prefixed with one UtxoState that contains the combined state of the removed UtxoStates.

pointLessThanTip :: Point -> Tip -> Bool

Is the given point earlier than the provided tip. Yes, if the point is the genersis point, no if the tip is the genesis point, otherwise, just compare the slots.

initialStatus :: OnChainTx -> TxStatus

The TxStatus of a transaction right after it was added to the chain

increaseDepth :: TxStatus -> TxStatus

Increase the depth of a tentatively confirmed transaction

chainConstant :: Depth

The depth (in blocks) after which a transaction cannot be rolled back anymore

dropOlder :: Monoid a => BlockNumber -> UtxoIndex a -> UtxoIndex a

Drop everything older than BlockNumber in the index.

transactionStatus :: BlockNumber -> TxIdState -> TxId -> Either TxStatusFailure TxStatus

Given the current block, compute the status for the given transaction by checking to see if it has been deleted.

transactionOutputStatus

Arguments

:: BlockNumber
Current block number for inspecting the state of the transaction output

-> TxIdState
Information on the state of a transaction. Needed for determining its status.

-> TxOutBalance
Balance of spent and unspent transaction outputs.

-> TxOutRef
Target transaction output for inspecting its state.

-> Either TxStatusFailure TxOutStatus

Given the current block, compute the status for the given transaction output by getting the state of the transaction that produced it and checking if the output is spent or unspent.

transactionOutputState :: TxOutBalance -> TxOutRef -> Maybe TxOutState

unspentOutputs :: UtxoState TxOutBalance -> Set TxOutRef

The UTXO set

spentOutputs :: UtxoState TxOutBalance -> Set TxOutRef

The spent output set

type ChainIndexState = UtxoIndex TxUtxoBalance

getResumePoints :: Member (BeamEffect Sqlite) effs => Eff effs [ChainPoint]

handleQuery :: (Member (State ChainIndexState) effs, Member (BeamEffect Sqlite) effs, Member (Error ChainIndexError) effs, Member (LogMsg ChainIndexLog) effs) => ChainIndexQueryEffect ~> Eff effs

handleControl :: forall effs. (Member (State ChainIndexState) effs, Member (Reader Depth) effs, Member (BeamEffect Sqlite) effs, Member (Error ChainIndexError) effs, Member (LogMsg ChainIndexLog) effs) => ChainIndexControlEffect ~> Eff effs

restoreStateFromDb :: Member (BeamEffect Sqlite) effs => Eff effs ChainIndexState