Plutus.ChainIndex.ChainIndexError

Documentation

data ChainIndexError

Constructors

InsertionFailed InsertUtxoFailed

RollbackFailed RollbackFailed

ResumeNotSupported

QueryFailedNoTip
Query failed because the chain index does not have a tip (not synchronised with node)

BeamEffectError BeamError

data InsertUtxoFailed

UTXO state could not be inserted into the chain index

Constructors

DuplicateBlock Tip
Insertion failed as there was already a block with the given number

InsertUtxoNoTip
The _usTip field of the argument was 'Last Nothing'

data RollbackFailed

Reason why the rollback operation failed

Constructors

RollbackNoTip
Rollback failed because the utxo index had no tip (not synchronised)

TipMismatch
Unable to roll back to expectedTip because the tip at that position was different

Fields

foundTip :: Tip

targetPoint :: Point

OldPointNotFound Point
Unable to find the old tip