Plutus.ChainIndex.Emulator

Documentation

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

data ChainIndexTxOutputs

List of outputs of a transaction. There is only an optional collateral output if the transaction is invalid.

Constructors

InvalidTx (Maybe ChainIndexTxOut)
The transaction is invalid so there is maybe a collateral output.

ValidTx [ChainIndexTxOut]

data ChainIndexTxOut

Constructors

ChainIndexTxOut

Fields

citoAddress :: CardanoAddress

We can't use AddressInAnyEra here because of missing FromJson instance for Byron era

citoValue :: Value

citoDatum :: OutputDatum

citoRefScript :: ReferenceScript

data ReferenceScript

Constructors

ReferenceScriptNone

ReferenceScriptInAnyLang ScriptInAnyLang

fromReferenceScript :: ReferenceScript -> Maybe (Versioned Script)

data ChainIndexTx

Constructors

ChainIndexTx

Fields

_citxTxId :: TxId

The id of this transaction.

_citxInputs :: [TxIn]

The inputs to this transaction.

_citxOutputs :: ChainIndexTxOutputs

The outputs of this transaction, ordered so they can be referenced by index.

_citxValidRange :: !SlotRange

The SlotRange during which this transaction may be validated.

_citxData :: Map DatumHash Datum

Datum objects recorded on this transaction.

_citxRedeemers :: Redeemers

Redeemers of the minting scripts.

_citxScripts :: Map ScriptHash (Versioned Script)

The scripts (validator, stake validator or minting) part of cardano tx.

_citxCardanoTx :: Maybe SomeCardanoApiTx

The full Cardano API tx which was used to populate the rest of the ChainIndexTx fields. Useful because ChainIndexTx doesn't have all the details of the tx, so we keep it as a safety net. Might be Nothing if we are in the emulator.

_InvalidTx :: Prism' ChainIndexTxOutputs (Maybe ChainIndexTxOut)

_ValidTx :: Prism' ChainIndexTxOutputs [ChainIndexTxOut]

data Tip

The tip of the chain index.

Constructors

TipAtGenesis

Tip

Fields

tipSlot :: Slot

Last slot

tipBlockId :: BlockId

Last block ID

tipBlockNo :: BlockNumber

Last block number

newtype BlockNumber

Constructors

BlockNumber

Fields

unBlockNumber :: Word64

citxCardanoTx :: Lens' ChainIndexTx (Maybe SomeCardanoApiTx)

citxData :: Lens' ChainIndexTx (Map DatumHash Datum)

citxInputs :: Lens' ChainIndexTx [TxIn]

citxOutputs :: Lens' ChainIndexTx ChainIndexTxOutputs

citxRedeemers :: Lens' ChainIndexTx Redeemers

citxScripts :: Lens' ChainIndexTx (Map ScriptHash (Versioned Script))

citxTxId :: Lens' ChainIndexTx TxId

citxValidRange :: Lens' ChainIndexTx SlotRange

blockId :: Block -> BlockId

Compute a hash of the block's contents.

data Point

When performing a rollback the chain sync protocol does not provide a block number where to resume from.

Constructors

PointAtGenesis

Point

Fields

pointSlot :: Slot

Slot number

pointBlockId :: BlockId

Block number

data TxOutBalance

The effect of a transaction (or a number of them) on the tx output set.

Constructors

TxOutBalance

Fields

_tobUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tobSpentOutputs :: Map TxOutRef TxId

Outputs spent by the transaction(s) along with the tx id that spent it

data TxConfirmedState

Constructors

TxConfirmedState

Fields

timesConfirmed :: Sum Int

blockAdded :: Last BlockNumber

validity :: Last TxValidity

data TxIdState

Constructors

TxIdState

Fields

txnsConfirmed :: Map TxId TxConfirmedState

Number of times this transaction has been added as well as other necessary metadata.

txnsDeleted :: Map TxId (Sum Int)

Number of times this transaction has been deleted.

data TxStatusFailure

Datatype returned when we couldn't get the state of a tx or a tx output.

Constructors

TxIdStateInvalid BlockNumber TxId TxIdState
We couldn't return the status because the TxIdState was in a ... state ... that we didn't know how to decode in transactionStatus.

TxOutBalanceStateInvalid BlockNumber TxOutRef TxOutBalance
We couldn't return the status because the TxOutBalance does not contain the target tx output.

InvalidRollbackAttempt BlockNumber TxId TxIdState

data Diagnostics

Constructors

Diagnostics

Fields

numTransactions :: Integer

numScripts :: Integer

numAddresses :: Integer

numAssetClasses :: Integer

numUnspentOutputs :: Int

numUnmatchedInputs :: Int

someTransactions :: [TxId]

unspentTxOuts :: [ChainIndexTxOut]

data TxOutState

Constructors

Spent TxId

Unspent

type TxOutStatus = RollbackState TxOutState

data RollbackState a

The rollback state of a Cardano transaction

Constructors

Unknown
The transaction is not on the chain. That's all we can say.

TentativelyConfirmed Depth TxValidity a
The transaction is on the chain, n blocks deep. It can still be rolled back.

Committed TxValidity a
The transaction is on the chain. It cannot be rolled back anymore.

type TxStatus = RollbackState ()

The status of a Cardano transaction

newtype Depth

How many blocks deep the tx is on the chain

Constructors

Depth

Fields

unDepth :: Int

data TxValidity

Validity of a transaction that has been added to the ledger

Constructors

TxValid

TxInvalid

UnknownValidity

_PointAtGenesis :: Prism' Point ()

_Point :: Prism' Point (Slot, BlockId)

tipAsPoint :: Tip -> Point

pointsToTip :: Point -> Tip -> Bool

txOutStatusTxOutState :: TxOutStatus -> Maybe TxOutState

Maybe extract the TxOutState (Spent or Unspent) of a TxOutStatus.

liftTxOutStatus :: TxOutStatus -> TxStatus

Converts a TxOutStatus to a TxStatus. Possible since a transaction output belongs to a transaction.

Note, however, that we can't convert a TxStatus to a TxOutStatus.

data TxUtxoBalance

The effect of a transaction (or a number of them) on the utxo set.

Constructors

TxUtxoBalance

Fields

_tubUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tubUnmatchedSpentInputs :: Set TxOutRef

Outputs spent by the transaction(s) that have no matching unspent output

tobSpentOutputs :: Lens' TxOutBalance (Map TxOutRef TxId)

tobUnspentOutputs :: Lens' TxOutBalance (Set TxOutRef)

data ChainSyncBlock

A block of transactions to be synced.

Constructors

Block

Fields

blockTip :: Tip

blockTxs :: [(ChainIndexTx, TxProcessOption)]

newtype TxProcessOption

User-customizable options to process a transaction. See #73 for more motivations.

Constructors

TxProcessOption

Fields

tpoStoreTx :: Bool

Should the chain index store this transaction or not. If not, only handle the UTXOs. This, for example, allows applications to skip unwanted pre-Alonzo transactions.

tubUnmatchedSpentInputs :: Lens' TxUtxoBalance (Set TxOutRef)

tubUnspentOutputs :: Lens' TxUtxoBalance (Set TxOutRef)

data RollbackFailed

Reason why the rollback operation failed

Constructors

RollbackNoTip
Rollback failed because the utxo index had no tip (not synchronised)

TipMismatch
Unable to roll back to expectedTip because the tip at that position was different

Fields

foundTip :: Tip

targetPoint :: Point

OldPointNotFound Point
Unable to find the old tip

data InsertUtxoFailed

UTXO state could not be inserted into the chain index

Constructors

DuplicateBlock Tip
Insertion failed as there was already a block with the given number

InsertUtxoNoTip
The _usTip field of the argument was 'Last Nothing'

data ChainIndexError

Constructors

InsertionFailed InsertUtxoFailed

RollbackFailed RollbackFailed

ResumeNotSupported

QueryFailedNoTip
Query failed because the chain index does not have a tip (not synchronised with node)

BeamEffectError BeamError

data CredentialMap

Set of transaction output references for each address.

data AssetClassMap

Set of transaction output references for each asset class.

unCredentialMap :: Iso' CredentialMap (Map Credential (Set TxOutRef))

data DiskState

Data that we keep on disk. (This type is used for testing only - we need other structures for the disk-backed storage)

unAssetClassMap :: Iso' AssetClassMap (Map AssetClass (Set TxOutRef))

addressMap :: Lens' DiskState CredentialMap

assetClassMap :: Lens' DiskState AssetClassMap

dataMap :: Lens' DiskState (Map DatumHash Datum)

redeemerMap :: Lens' DiskState (Map RedeemerHash Redeemer)

scriptMap :: Lens' DiskState (Map ScriptHash (Versioned Script))

txMap :: Lens' DiskState (Map TxId ChainIndexTx)

diagnostics :: DiskState -> Diagnostics

data ChainIndexQueryEffect r where

Constructors

DatumFromHash :: DatumHash -> ChainIndexQueryEffect (Maybe Datum)
Get the datum from a datum hash (if available)

ValidatorFromHash :: ValidatorHash -> ChainIndexQueryEffect (Maybe (Versioned Validator))
Get the validator from a validator hash (if available)

MintingPolicyFromHash :: MintingPolicyHash -> ChainIndexQueryEffect (Maybe (Versioned MintingPolicy))
Get the monetary policy from an MPS hash (if available)

RedeemerFromHash :: RedeemerHash -> ChainIndexQueryEffect (Maybe Redeemer)
Get the redeemer from a redeemer hash (if available)

StakeValidatorFromHash :: StakeValidatorHash -> ChainIndexQueryEffect (Maybe (Versioned StakeValidator))
Get the stake validator from a stake validator hash (if available)

UnspentTxOutFromRef :: TxOutRef -> ChainIndexQueryEffect (Maybe DecoratedTxOut)
Get the TxOut from a TxOutRef (if available)

TxOutFromRef :: TxOutRef -> ChainIndexQueryEffect (Maybe DecoratedTxOut)
Get the TxOut from a TxOutRef (if available)

TxFromTxId :: TxId -> ChainIndexQueryEffect (Maybe ChainIndexTx)
Get the transaction for a tx ID

UtxoSetMembership :: TxOutRef -> ChainIndexQueryEffect IsUtxoResponse
Whether a tx output is part of the UTXO set

UtxoSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect UtxosResponse
Unspent outputs located at addresses with the given credential.

UnspentTxOutSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect (QueryResponse [(TxOutRef, DecoratedTxOut)])
Get the unspent txouts located at an address This is to avoid multiple queries from chain-index when using utxosAt

DatumsAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect (QueryResponse [Datum])
get the datums located at addresses with the given credential.

UtxoSetWithCurrency :: PageQuery TxOutRef -> AssetClass -> ChainIndexQueryEffect UtxosResponse
Unspent outputs containing a specific currency (AssetClass).

Note that requesting unspent outputs containing Ada should not return anything, as this request will always return all unspent outputs.

TxsFromTxIds :: [TxId] -> ChainIndexQueryEffect [ChainIndexTx]
Get the transactions for a list of tx IDs.

TxoSetAtAddress :: PageQuery TxOutRef -> Credential -> ChainIndexQueryEffect TxosResponse
Outputs located at addresses with the given credential.

GetTip :: ChainIndexQueryEffect Tip
Get the tip of the chain index

data ChainIndexControlEffect r where

Constructors

AppendBlocks :: [ChainSyncBlock] -> ChainIndexControlEffect ()
Add new blocks to the chain index.

Rollback :: Point -> ChainIndexControlEffect ()
Roll back to a previous state (previous tip)

ResumeSync :: Point -> ChainIndexControlEffect ()
Resume syncing from a certain point

CollectGarbage :: ChainIndexControlEffect ()
Delete all data that is not covered by current UTxOs.

GetDiagnostics :: ChainIndexControlEffect Diagnostics

datumFromHash :: forall effs. Member ChainIndexQueryEffect effs => DatumHash -> Eff effs (Maybe Datum)

validatorFromHash :: forall effs. Member ChainIndexQueryEffect effs => ValidatorHash -> Eff effs (Maybe (Versioned Validator))

mintingPolicyFromHash :: forall effs. Member ChainIndexQueryEffect effs => MintingPolicyHash -> Eff effs (Maybe (Versioned MintingPolicy))

redeemerFromHash :: forall effs. Member ChainIndexQueryEffect effs => RedeemerHash -> Eff effs (Maybe Redeemer)

stakeValidatorFromHash :: forall effs. Member ChainIndexQueryEffect effs => StakeValidatorHash -> Eff effs (Maybe (Versioned StakeValidator))

unspentTxOutFromRef :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs (Maybe DecoratedTxOut)

txOutFromRef :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs (Maybe DecoratedTxOut)

txFromTxId :: forall effs. Member ChainIndexQueryEffect effs => TxId -> Eff effs (Maybe ChainIndexTx)

utxoSetMembership :: forall effs. Member ChainIndexQueryEffect effs => TxOutRef -> Eff effs IsUtxoResponse

utxoSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs UtxosResponse

unspentTxOutSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs (QueryResponse [(TxOutRef, DecoratedTxOut)])

datumsAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs (QueryResponse [Datum])

utxoSetWithCurrency :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> AssetClass -> Eff effs UtxosResponse

txsFromTxIds :: forall effs. Member ChainIndexQueryEffect effs => [TxId] -> Eff effs [ChainIndexTx]

txoSetAtAddress :: forall effs. Member ChainIndexQueryEffect effs => PageQuery TxOutRef -> Credential -> Eff effs TxosResponse

getTip :: forall effs. Member ChainIndexQueryEffect effs => Eff effs Tip

appendBlocks :: forall effs. Member ChainIndexControlEffect effs => [ChainSyncBlock] -> Eff effs ()

rollback :: forall effs. Member ChainIndexControlEffect effs => Point -> Eff effs ()

resumeSync :: forall effs. Member ChainIndexControlEffect effs => Point -> Eff effs ()

collectGarbage :: forall effs. Member ChainIndexControlEffect effs => Eff effs ()

getDiagnostics :: forall effs. Member ChainIndexControlEffect effs => Eff effs Diagnostics

data InsertUtxoPosition

Outcome of inserting a UtxoState into the utxo index

Constructors

InsertAtEnd
The utxo state was added to the end. Returns the new index

InsertBeforeEnd
The utxo state was added somewhere before the end. Returns the new index and the tip

data ChainIndexLog

Constructors

InsertionSuccess Tip InsertUtxoPosition

ConversionFailed FromCardanoError

RollbackSuccess Tip

Err ChainIndexError

TxNotFound TxId

TxOutNotFound TxOutRef

TipIsGenesis

NoDatumScriptAddr ChainIndexTxOut

BeamLogItem BeamLog

data ChainIndexEmulatorState

Constructors

ChainIndexEmulatorState

Fields

_diskState :: DiskState

_utxoIndex :: UtxoIndex TxUtxoBalance

diskState :: Lens' ChainIndexEmulatorState DiskState

utxoIndex :: Lens' ChainIndexEmulatorState (UtxoIndex TxUtxoBalance)

handleQuery :: forall effs. (Member (State ChainIndexEmulatorState) effs, Member (Error ChainIndexError) effs, Member (LogMsg ChainIndexLog) effs) => ChainIndexQueryEffect ~> Eff effs

handleControl :: forall effs. (Member (State ChainIndexEmulatorState) effs, Member (Error ChainIndexError) effs, Member (LogMsg ChainIndexLog) effs) => ChainIndexControlEffect ~> Eff effs

serveChainIndex :: forall effs. (Member (Error ServerError) effs, Member ChainIndexQueryEffect effs, Member ChainIndexControlEffect effs) => ServerT API (Eff effs)

serveChainIndexQueryServer

Arguments

:: Int
Port

-> TVar ChainIndexEmulatorState
Chain index state (TODO: When the disk state is stored in a DB, replace this with DB connection info)

-> IO ()