Plutus.ChainIndex.Types

Description

Misc. types used in this package

Documentation

data ChainIndexTx

Constructors

ChainIndexTx

Fields

_citxTxId :: TxId

The id of this transaction.

_citxInputs :: [TxIn]

The inputs to this transaction.

_citxOutputs :: ChainIndexTxOutputs

The outputs of this transaction, ordered so they can be referenced by index.

_citxValidRange :: !SlotRange

The SlotRange during which this transaction may be validated.

_citxData :: Map DatumHash Datum

Datum objects recorded on this transaction.

_citxRedeemers :: Redeemers

Redeemers of the minting scripts.

_citxScripts :: Map ScriptHash (Versioned Script)

The scripts (validator, stake validator or minting) part of cardano tx.

_citxCardanoTx :: Maybe SomeCardanoApiTx

The full Cardano API tx which was used to populate the rest of the ChainIndexTx fields. Useful because ChainIndexTx doesn't have all the details of the tx, so we keep it as a safety net. Might be Nothing if we are in the emulator.

data ChainIndexTxOutputs

List of outputs of a transaction. There is only an optional collateral output if the transaction is invalid.

Constructors

InvalidTx (Maybe ChainIndexTxOut)
The transaction is invalid so there is maybe a collateral output.

ValidTx [ChainIndexTxOut]

data ChainIndexTxOut

Constructors

ChainIndexTxOut

Fields

citoAddress :: CardanoAddress

We can't use AddressInAnyEra here because of missing FromJson instance for Byron era

citoValue :: Value

citoDatum :: OutputDatum

citoRefScript :: ReferenceScript

data ReferenceScript

Constructors

ReferenceScriptNone

ReferenceScriptInAnyLang ScriptInAnyLang

newtype BlockId

Block identifier (usually a hash)

Constructors

BlockId

Fields

getBlockId :: ByteString

blockId :: Block -> BlockId

Compute a hash of the block's contents.

data Tip

The tip of the chain index.

Constructors

TipAtGenesis

Tip

Fields

tipSlot :: Slot

Last slot

tipBlockId :: BlockId

Last block ID

tipBlockNo :: BlockNumber

Last block number

data Point

When performing a rollback the chain sync protocol does not provide a block number where to resume from.

Constructors

PointAtGenesis

Point

Fields

pointSlot :: Slot

Slot number

pointBlockId :: BlockId

Block number

pointsToTip :: Point -> Tip -> Bool

tipAsPoint :: Tip -> Point

_PointAtGenesis :: Prism' Point ()

_Point :: Prism' Point (Slot, BlockId)

data TxValidity

Validity of a transaction that has been added to the ledger

Constructors

TxValid

TxInvalid

UnknownValidity

type TxStatus = RollbackState ()

The status of a Cardano transaction

type TxOutStatus = RollbackState TxOutState

data RollbackState a

The rollback state of a Cardano transaction

Constructors

Unknown
The transaction is not on the chain. That's all we can say.

TentativelyConfirmed Depth TxValidity a
The transaction is on the chain, n blocks deep. It can still be rolled back.

Committed TxValidity a
The transaction is on the chain. It cannot be rolled back anymore.

data TxOutState

Constructors

Spent TxId

Unspent

liftTxOutStatus :: TxOutStatus -> TxStatus

Converts a TxOutStatus to a TxStatus. Possible since a transaction output belongs to a transaction.

Note, however, that we can't convert a TxStatus to a TxOutStatus.

txOutStatusTxOutState :: TxOutStatus -> Maybe TxOutState

Maybe extract the TxOutState (Spent or Unspent) of a TxOutStatus.

newtype BlockNumber

Constructors

BlockNumber

Fields

unBlockNumber :: Word64

newtype Depth

How many blocks deep the tx is on the chain

Constructors

Depth

Fields

unDepth :: Int

data Diagnostics

Constructors

Diagnostics

Fields

numTransactions :: Integer

numScripts :: Integer

numAddresses :: Integer

numAssetClasses :: Integer

numUnspentOutputs :: Int

numUnmatchedInputs :: Int

someTransactions :: [TxId]

unspentTxOuts :: [ChainIndexTxOut]

data TxConfirmedState

Constructors

TxConfirmedState

Fields

timesConfirmed :: Sum Int

blockAdded :: Last BlockNumber

validity :: Last TxValidity

data TxStatusFailure

Datatype returned when we couldn't get the state of a tx or a tx output.

Constructors

TxIdStateInvalid BlockNumber TxId TxIdState
We couldn't return the status because the TxIdState was in a ... state ... that we didn't know how to decode in transactionStatus.

TxOutBalanceStateInvalid BlockNumber TxOutRef TxOutBalance
We couldn't return the status because the TxOutBalance does not contain the target tx output.

InvalidRollbackAttempt BlockNumber TxId TxIdState

data TxIdState

Constructors

TxIdState

Fields

txnsConfirmed :: Map TxId TxConfirmedState

Number of times this transaction has been added as well as other necessary metadata.

txnsDeleted :: Map TxId (Sum Int)

Number of times this transaction has been deleted.

data TxUtxoBalance

The effect of a transaction (or a number of them) on the utxo set.

Constructors

TxUtxoBalance

Fields

_tubUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tubUnmatchedSpentInputs :: Set TxOutRef

Outputs spent by the transaction(s) that have no matching unspent output

tubUnspentOutputs :: Lens' TxUtxoBalance (Set TxOutRef)

tubUnmatchedSpentInputs :: Lens' TxUtxoBalance (Set TxOutRef)

data TxOutBalance

The effect of a transaction (or a number of them) on the tx output set.

Constructors

TxOutBalance

Fields

_tobUnspentOutputs :: Set TxOutRef

Outputs newly added by the transaction(s)

_tobSpentOutputs :: Map TxOutRef TxId

Outputs spent by the transaction(s) along with the tx id that spent it

tobUnspentOutputs :: Lens' TxOutBalance (Set TxOutRef)

tobSpentOutputs :: Lens' TxOutBalance (Map TxOutRef TxId)

data ChainSyncBlock

A block of transactions to be synced.

Constructors

Block

Fields

blockTip :: Tip

blockTxs :: [(ChainIndexTx, TxProcessOption)]

newtype TxProcessOption

User-customizable options to process a transaction. See #73 for more motivations.

Constructors

TxProcessOption

Fields

tpoStoreTx :: Bool

Should the chain index store this transaction or not. If not, only handle the UTXOs. This, for example, allows applications to skip unwanted pre-Alonzo transactions.

Lenses

citxTxId :: Lens' ChainIndexTx TxId

citxInputs :: Lens' ChainIndexTx [TxIn]

citxOutputs :: Lens' ChainIndexTx ChainIndexTxOutputs

citxValidRange :: Lens' ChainIndexTx SlotRange

citxData :: Lens' ChainIndexTx (Map DatumHash Datum)

citxRedeemers :: Lens' ChainIndexTx Redeemers

citxScripts :: Lens' ChainIndexTx (Map ScriptHash (Versioned Script))

citxCardanoTx :: Lens' ChainIndexTx (Maybe SomeCardanoApiTx)

_InvalidTx :: Prism' ChainIndexTxOutputs (Maybe ChainIndexTxOut)

_ValidTx :: Prism' ChainIndexTxOutputs [ChainIndexTxOut]

fromReferenceScript :: ReferenceScript -> Maybe (Versioned Script)

Orphan instances

ToSchema ScriptInAnyLang

Instance details

Methods

declareNamedSchema :: Proxy ScriptInAnyLang -> Declare (Definitions Schema) NamedSchema

Serialise ScriptInAnyLang

Instance details

Methods

encode :: ScriptInAnyLang -> Encoding

decode :: Decoder s ScriptInAnyLang

encodeList :: [ScriptInAnyLang] -> Encoding

decodeList :: Decoder s [ScriptInAnyLang]